requirements.txt
# AI Radio - Requirements
# Python 3.11+ required

# Core dependencies (existing)
PyYAML>=6.0.1
asyncio>=3.4.3

# GPIO for Raspberry Pi (optional, for hardware)
gpiozero>=2.0.1
RPi.GPIO>=0.7.1

# Audio processing
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
scipy>=1.11.0

# Audio effects and processing
librosa>=0.10.0
pedalboard>=0.9.0  # Spotify's audio effects library

# LLM Client Libraries
# OpenAI
openai>=1.12.0

# Anthropic Claude
anthropic>=0.18.0

# Ollama (local LLM)
ollama>=0.1.6

# LlamaCPP (local LLM alternative)
llama-cpp-python>=0.2.0

# Text-to-Speech Engines
# Piper TTS (recommended - fast, local, high quality)
piper-tts>=1.2.0

# Coqui TTS (alternative - more voices but slower)
TTS>=0.22.0

# eSpeak (lightweight fallback)
py-espeak-ng>=0.1.8

# HTTP client for API calls
httpx>=0.26.0
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Audio file handling
mutagen>=1.47.0  # Audio metadata

# Signal processing for sonification
# (covered by numpy, scipy above)

# Utilities
asyncio-throttle>=1.0.0  # Rate limiting for API calls
cachetools>=5.3.0  # Caching

# CLI and logging
colorama>=0.4.6  # Colored terminal output
python-json-logger>=2.0.7

# Optional: For advanced audio synthesis
# mido>=1.3.0  # MIDI processing (for musical sonification)

# Testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.9.0
